Renamed hasMove to hasMark
Since you are marking the squares and it returns true if the square was
marked.  To me if it has a Move, I don't know at first if it meant it
had a Move available or a Move already made.  Plus move feels more like
a chess term where you are moving pieces around.  Anyway, small rename
for readability. No hard feelings is the mark as verb, mark as noun
issue is worse naming wise for some folks.