[Initialization] Decouple script loading and environment setup from the root view
Introduces RCTRootViewController, which handles the controller-appropriate work that RCTRootView was doing. This decouples the view from the script loading and JS environment setup (w/the bridge), which helps ReactKit be less of a framework and more of a library.

The example apps have been updated and work. On the Facebook side you will need to replace RCTRootView with RCTRootViewController which should be fairly mechanical with VC containment.