[React-Native] Update jstransform/esprima
Summary: This updates jstransform and esprima so that Relay/Dlite can be updated in a subsequent diff. It is too big to send to Phabricator so @jingc manually reviewed this diff with me. @jeffmo also took a look at the manually changed parts of this diff.

Here is a paste of the manual changes: https://phabricator.fb.com/P19736901

After changing `package.json` I ran an `npm update` that also pulled in some other npm changes. I've tested everything and it seems fine.

Test Plan:

* `./runJestTests.sh` --reset-cache all pass
* `./runServerHere.sh` --reset-cache` works; the Catalyst shell and the Groups app work as expected

CC:

Task ID: #

Blame Rev: