[react-packager][streamline oss] Move open sourced JS source to react-native-github
Summary:
In an effort to make the open source repo match our repo, this moves all the
open sourced JS files to a dedicated directory `FBReactKit/js/react-native-github`
that exactly matches github.  Duplicate files were removed by using `mv`
instead of `cp` in the oss_export script (which is also updated). The rest of
the work was mostly in testConfig.json, server.js, and runServer.js. There were
a few other tweaks along the way, like not using relative includes in Shrubbery
(not sure why it was an issue since it works in Movies, but oh well).  I also
added the website and docs.

Test Plan:
* test the internal shell using the normal script `./runServerHere.sh`
* cp to github and test the open source app by opening the movies app and starting the server using `npm start`
* test chrome debugging and click to line stack traces.