[react-packager] Cleanup option passing and validation
Summary:
This is the second and final diff of for cleaning up options passing and validation.
This diff adds a new library module called `declareOpts` that's used to declare the set of options, their types, and if their required, and their default values.
declareOpts is then used by every module in the program to declare it's options.

Test Plan:
* `cd JSAppServer && jest`
* start server and make sure the apps work.