Only update Redbox when offscreen
Summary:
Because the React application isn't terminated by a redbox error, it will often produce multiple
errors in sequence, with the later ones replacing the earlier ones in the redbox trace.

However, for debugging purposes, you almost always want to know what the *first* error was, not
what some knock-on effect later on was. For example, any error in your JavaScript will eventually
result in the module failing to be resgistered, so the only error you ever see in the Redbox as a
result of JS coding error is "module was not registered" instead of the actual cause.

This diff changes the Redbox behaviour so that the display isn't updated if a new error comes
in when the redbox is already on screen. This makes it more useful as a debugging tool.

Test Plan: * introduce a deliberate error in a .js file, such as reference to a non-existant property. Observe that the error is reported in the redbox.