[ListView] Make LVDS.getSectionLengths() be indexed by section ID
Currently ListViewDataSource.getSectionLengths() returns an array of section lengths, where the nth element of the array is the number of rows in the nth section. There is a tiny bit of an API mismatch since elsewhere section IDs, which are strings, are primarily used. This diff changes getSectionLengths to return an object whose keys are section IDs and values are the number of rows in each respective section.

Before: `[10, 20, 30]`
After: `{firstSection: 10, secondSection: 20, thirdSection: 30}`