[MAdMan][Android] Make things look more Androidy
Summary:
Fudge navigation bar styles to look more Android-y and use Android-specific assets.

This isn't by any means a final state, really just a hack for demo purposes. Going forward we will actually be using the Android toolbar in the navigation bar. However, this is still a useful step towards that because of the assets, color definitions, etc.

Test Plan:
{F22019741}
{F22019738}
{F22019737}