[ReactNative] unpod Movies and TicTacToe examples
Summary:
We're going to come up with something better (probably use the xcodeproj project
to create them in a more configurable and easily replaced with Buck way).  Keeps
the podspecs on libraries so it's easy to use them if desired, but the xcodeproj
files are just manual now.

Test Plan: Open xamples, clean build runs them all fine.