[ReactNative] Move merge & mergeInto from downstream to vendor
Summary:
Previously, those files were taken from React but it now throws a warning when you just include them. So I removed them from react, but now they are not in the open source repo anymore. Moving them to vendor/

Note that React is going to get rid of them in the next update, so this is good for long-term

Test Plan: Run app launcher and madman, it works