[react-packager] Add uglify-js library
Summary: We use the uglify-js library to minify the bundle, however, we didn't have it in the package.json. This adds it to package.json and updates the library. I also added to the oss package.json because in an upcoming diff react-packager will take care of minification.

Test Plan:
* ./runJestTests.sh
* test bundle creation with bundle.sh
* test load_dependencies.js script
* start the server and click around shell app