[react-packager] Cleanup package.json
Summary:
After the unforking of the JavaScript we left two package.json that were duplicated and needed to stay in sync:
* `js/package.json`
* `js/react-native-github/package.json`

There are a few problems with this:
* We may update one and forget to update the other
* This was causing weirdness with jest tests. It sees a package.json but can only find the dependencies in the parent directory. Required some hacks to get around
* We have a lot of crap that we don't need

In this diff, I make sure react-native-github defines it's own packages and doesn't rely on the parent package to support it. In a following diff I'll check in node_modules for react-native-github, and that's why I added a line in the export script to delete node_modules.

Test Plan:
*./runJestTests.sh
*test bundle creation with bundle.sh
*test load_dependencies.js script
*start the server and click around shell app