[ReactNative] Use strings instead of constants for keyboardDismissMode
Summary:
On a very long thread that I unfortunately can't find back, @jwalke got convinced that we could actually type-check values as string. This is a much better API for the user as he doesn't have to type a crazy long constant and is consistent with styles.

I'm happy with the user-facing API but we're probably going to want to do something else for the actual implementation. In any case, we need to future-proof the API for open source, we can change the internals at will.

Test Plan: Run madman, play around the three different values and wrong values. Make sure that the proper numbers are sent to obj-c.