[ReactNative] Github repo's gitignore is written at export
Summary:
With checking in node_modules into github-react-native	it should be tracked by	source control. However	we have	a .gitignore inside the repository that in reality belongs to the open source repo.

This adds a bit	    of complexity to the export script by having it	write out .gitignore file, but I think it's worth it to have a clean seperation.

Test Plan:
* ./runJestTests.sh
* run `oss_export.sh` and make sure the	export succeeds	and is correct.