[ReactNative] Use spread operator and .propTypes for ScrollView/ListView
Summary:
It turns out that propTypes is automatically added to the component class, so don't need to manually export it. Also, we can use spread operator instead of merge.

Those two changes make the code easier and work with @fkling source extraction :)

Test Plan: Changed the attributes of ListView and ScrollView and make sure that propTypes are still applied