[react-packager] Start converting options to query params
Summary: we're using this weird `.` seperator in the url for options. This changes options to be query params, however, maintains backwards compatibility to not break any clients. In the future any options should be added as query params. (In an upcoming diff I actually add dev/minify as query param options).

Test Plan:
* runJestTests.sh
* open shell app and click around