[react-packager] onchange endpoint that informs of changes
Summary:
our current solution for live reloading is pulling the server which is spammy and not that great. This exposes a long-polling end point that will end the request with a message once a file change.

Right now it's really simple, but if you want more information on what change we can do that.

Test Plan:
* ./runJestTests.sh
* open browser to http://localhost:8081/onchange
* change a file
* observer message in browser {changed: true}