[React Native] Update core modules for React 0.13
Summary:
This fixes up the React Native modules to work correctly with React 0.13. There were a bunch of internal refactors and I also fixed up a couple of components that warned or broke (prop mutation, `._currentElement` gone, plain functions in JSX don't work).

I don't know of any outstanding problems with this diff, though we may want to change the native components to have a composite wrapper, like we do for native (DOM) components on the web? I don't know the right place to hook that in though.

Depends on D1893948.

Test Plan: Clicked around in Madman a lot. Tested that Madman navigator animations still feel smooth on a 4S. Clicked around in Groups as well, and in a few other apps within the Catalyst shell like the UI Explorer.