[ReactNative] Move around and reformat comments for the documentation
Summary:
react-docgen expects the comment to be right above the react declaration to be picked up, moved it there. (I actually think that this is a good change)

Also replaced the indent code by space to ``` in order to syntax highlight on the markdown parser

Test Plan: Run all the docs and make sure that the comments appear.