[ReactNative] Unify ScrollView.android and ScrollView.ios
Summary:
ScrollView today is split between two files, the only meaningful difference between the two is chosing which native implementation to use. This diff unifies them using a if check on Platform.

The original motivation behind this change is to get rid of ScrollViewPropTypes module which doesn't work with the propTypes extraction tool. It turns out that Android version doesn't have any comments on propTypes either.

In the future, we'll probably want to have a <ScrollView> which is the lowest common denominator between both platforms and then <ScrollViewIOS> and <ScrollViewAndroid> that contains all the special attributes (like all the bouncy attributes for iOS).

Test Plan:
Run madman on iOS, make sure it scrolls and no related warnings
Copy the js files in fbandroid, run madman on Android, make sure it works and no related warnings