[ReactNative] Remove nativePropTypes
Summary:
The motivation behind removing it is that it doesn't work with the prop types parser. It has also never been pushed outside of ScrollView.

We'll probably want to revisit it in the future as I think it's a good idea, but right now it's easier to remove it.

Test Plan:
Run app launcher (onScroll & treeshold)
Run @scale (horizontal pagination)
Run scrollview example
Run madman

Make sure all of them work fine