[ReactNative] allowLossyConversion on NSString->NSData conversion
Summary:
The result from an execJSCall could contain a string that can't
be converted to NSData without being lossy. This causes
dataUsingEncoding: to return nil which causes a crash when we pass it to
NSJSONSerialization.

See
https://developer.apple.com/library/mac/documentation/Cocoa/Reference/Foundation/Classes/NSString_Class/#//apple_ref/occ/instm/NSString/dataUsingEncoding:allowLossyConversion:

"An NSData object containing a representation of the receiver encoded
using encoding. Returns nil if flag is NO and the receiver can’t be
converted without losing some information (such as accents or case)."

Test Plan:
I'd love suggestions for a way to test that this is actually
the issue we're facing.