Unforked RKNavigator, RKScrollView and RKView
Summary:
This diff unforks RKNavigator, RKScrollView and RKView, and a few other bits and pieces

NOTE:

I've moved getScrollViewContentSize() out of RKUIManager and into the RCTScrollViewManager. I've updated all the call sites in JS, and it doesn't seem to be used in MAdMan - not sure of there are any Android changes required?

Test Plan: Test Groups, MAdMan, Catalyst, etc