[react-packager] package.json cleanup (seperate packager into it's own package)
Summary:
We currently have react-native-github massive package.json that has all the dependencies of react-native product and
core code and the packager. This meant that we had to have an ever-expanding blacklist. We also had a lot of duplications in our package.json(s). This diff cleans all that out and make sure the packager is it's own package.

Test Plan:
    * runJestTests
    * start server and click around in the shell app
    * export oss package and start example app