Fixed sticky section headers in ListView
Summary:
The sticky headers in ListView weren't working correctly, as reported here: https://github.com/facebook/react-native/issues/141

It turns out that this was because the property was only ever being set to [1] (an array with a single index value of 1, reprenting the first header) so all the other headers were treated as regular rows and were scrolling under the first header.

It turns out that *that* was because the diffing function for stickyHeaders was wrong, or, more precisely, was a function when it should have been an object containing a function.

Test Plan: Run UIExplorer. Verify that paged scrolling example works.