[catalyst] set up directory specific rql transform
Summary: Splitting out from D1828349. Introduced a new `GraphQLSchemaPath` module that can be used to determine which graphql schema to use for rql transforms. This allows us to use the same server for files that require different s    chemas. `getVisitorList` memoizes the visitorLists based on schema paths to ensure that the added overhead is minimal and negligible.

Test Plan:
- `arc build`, build Catalyst and use Treehouse, rebase D1828349 on top and build mobile home
- `runJestTests.sh`
- sandcastle
- `ibgs` and follow call sites, ensure all calls to `transform` pass in `filePath`