[ReactNative] Fix RCTMapManager retaining cycle
Summary:
MapManager was calling the javascript callback before invalidating the timer,
which in some cases which cause a retaining cycle since the javascript would
update the map region.

Test Plan:
Run the MAdMan app, go the app, edit the location, close the map and then open
it again. The radius shouldn't be flickering anymore.