[ReactNative] Fix ScrollView.scrollTo()
Summary: This broke when we upgraded to the latest version of React which doesn't expose this._rootNodeID anymore. Instead, we need to use this.getNodeHandle().

Test Plan:
Run madman on iOS, scroll to the bottom of the landing page, tap twice on the first tab, make sure it goes to the top.

ibgs for _rootNodeID, make sure that it does work.