Fixed scrollOffset bug
Summary:
Certain properties, when set on scrollview, reset the scrollOffset to {0,0}.

In the old RKScrollView implementation, we explictly set the scrollOffset property last, so it wouldn't be overriden.

But in the new system, we can't control the order that properties are set, and this lead to https://our.intern.facebook.com/intern/tasks/?t=6467793

This diff fixes that by making it the responsibility of the property setters to restore the scroll offset.

Test Plan:
* Run Groups, open a multiphoto post and verify that tapping a photo > 1 doesn't reset to 1st image
* Smoke test UIExplorer, etc to check scrollviews still work, generally.