[ReactNative] Add integration test stuff
Summary:
This creates a new lib `RCTTest` with a test runner and a test module so JS tests
can tell native when they are done.  Uses these to implement a few basic tests
in a new `IntegrationTests` app.  The App can also run the tests like a normal
app, which makes debugging easier.

This also includes some additions to redbox so we can get errors, and root view
so we can inject modules.

Test Plan:
cmd+U - all tests pass.  Muck with the JS files and things fail as
expected.  cmd+R launches app which looks good - tapping the only test runs and
renders it.

{F21932131}