[React Native] Prevent crash in redbox code with two thrown errors
Summary:
Previously, if two errors were thrown then we would assign the first to `_lastStackTrace`, reload the table, then assign the second to `_lastStackTrace` without refreshing because we prefer to show the first error. This could result in `tableView:cellForRowAtIndexPath:` later crashing if the second stack trace is shorter and mixing the stack frames shown even if the stack traces are the same length, because we updated the data the table reads from without reloading the table.

We don't use these variables for anything else so I now assign them only when showing a new stack trace.

Test Plan: Throw two errors from JS, see the first in redbox, then scroll the stack trace and see the app not crash.