[ReactNative] Cleanup StyleSheet API
Summary:
We only want to expose StyleSheet.create and not the other validation functions. In this diff, I move all the validation functions to its own separate module called StyleSheetValidation, which is not exposed in require('react-native').

It also reformats the header to be markdown friendly

Test Plan: ./runJestTests.sh StyleSheet