[ReactNative] Allow single callbacks in NativeModules
Summary:
We discussed this as a team and decided there are defitely times where
we shouldn't require an error callback to go with a success callback (e.g.
`RKAsyncLocalStorage` passes an error value as the first arg to a single
callback, which is standard node convention).  This will let us switch all the
annoying `NativeModulesDeprecated` references, but I'll do that codemod in a
stacked diff.

Test Plan:
Switched `AsyncStorage` from `NativeModulesDeprecated` to
`NativeModules` - no more invariant violation, and everything still works great.