[ReactNative] Simplify and test interpolators
Summary: I was super bothered by the useless constants with values 0, 1, 1 and the fact that it was mutating values in the middle of the expression

Test Plan: I first generated many samples and added them to the tests. Then, I went ahead and codemod the code, making sure that at every step it still worked.