[RFC][ReactNative] increase timer resolution
Summary:
I realized reading our new docs that our timer resolution is only as
small as a frame, whereas browsers are much faster (a quick fiddle on mobile
safari shows setTimeout(fn, 0) will call within about 5 milliseconds).

This diff adds a quick check that immediately enqueues the timer callback if the
timeout is 0 - it's a little arbitrary, but I figured would cover the
common usage of setTimeout(fn, 0) to queue something up for the next event loop
as fast as possible.  This also makes a distinction between `setTimeout(fn, 0)`
and `requestAnimationFrame(fn)`, which adds functionality and is closer to
expectations, both inherant in the API and from working with browsers.

This may also be useful for up-coming time-slicing perf optimizations - we'll be
able to yield the JS thread to process pending gestures that are sitting in the
queue without wasting the rest of the frame.

It would of course be nicer if we could do something more accurate than
immediate or the next frame, but I'm not sure it's worth building the machinery
to make that happen.  It's also possible this may break some product
assumptions...

Test Plan:
Ran some simple timer resolution tests - we can call timers at a rate
of about 5000/second in sim, not sure what it will be on device.