[ReactNative][RFC] Bring Chrome debugger to OSS. Part 1
Summary:
This brings WebSocket based executor implementation to OSS ReactKit. Per discussion with
@grp and @a2 I'm just including SocketRocket implementation (h and m files) into ReactKit.

This diff is RFC, I want to make sure I'm moving in the right direction.

The next step would be copying websocket proxy stuff to OSS.

Test Plan: It compiles, but the debugger doesn't start because server side is not there yet.