[react-packager] change all but one `ix` to `require`
Summary:
This diff changes most ixs to requires. As mentioned before this work is to get us closer to a point were asset management is completely handled by the packager and is abstracted from the different platforms react-native (and web in the future) supports.

There was only one ix that we couldn't convert and that's part of the F8 app, and is appartently a hack to reduce app size. I renamed ix to ix_DEPRECATED so it doesn't get used going forward.

Test Plan:
* runJestTests.sh
* run server and click around apps
* export oss and run UIExplorer
* Make sure the image example works