[react-packager] Allow entry point extensions like .ios.js
Summary:
Because we used to pass options `.` delimeted in the url this made the packager think of anything with a dot in the name as an option. Which made it not work when the entry point file has a `.` in it. e.g. index.ios.js.

This diff hard codes backwards compat urls like `includeRequire` and `runModule` and leaves the rest as the filename.

As a bonus this fixes another problem by making runModule default to true.

Test Plan:
* runJestTests.sh
* create a file called index.ios.js
* start server
* request bundle with that file
* bundle is served