[ReactNative] Returning actual contentSize for RCTScrollViewManager
Summary:
Turns out only THGroupwView.js calls this so this change should only affect THGroupView.
##_contentSize## is never really set anywhere that I am aware of, so it would always be CGRectZero.
I am not entirely sure when this started happening. All other instances of contentSize seems to be based on the scrollView prop as well.

Maybe the more correct fix is to override how we return ##contentSize##. Not entirely sure though.

Test Plan: Scrolling in permalink mode is fixed. Insets are correct.