Work around flow bug with exports
Summary:
this file seems to be a perfect storm that confuses Flow to the point that it doesn't know the types of any of the modules exported by `react-native`, but does know that the keys exist, and does know the types of all of the stuff exported by `React`.

Exporting a spread of a require seems to be the problem, so I'm hacking around it by using `Object.create` to create a cloned object. This should be a very temporary workaround (hopefully we can fix the flow bug in time for next week's flow deploy, and then revert this), so I tried to keep the impact on blame minimal.

Test Plan: added @flow to `TimerMixin` and `PixelRatio`, and typed `PixelRatio.get` to return a string. before this, no errors. after this, it knows that `TimerMixin` isn't an actual `React.createMixin`, and it knows that `1 / PixelRatio.get()` is bogus.