Added a graph view and a ReactNative metric that displays current queue and execution time for the JS thread.
Summary: Changes the default "perf overlay" setting in catalyst to use FBPerformanceMonitor.  Adds a graph view option to FBPerformanceMonitor.  Adds a ReactNative metric which shows the execution and queueing time for javascript calls.

Test Plan: Run the perf overlay in the Groups app, MAdMan, and the Catalyst shell.  Make sure all the metrics are in the correct ballpark, and reset properly while browsing through the apps.