[ReactNative] Expanded license on js files
Summary:
We need to use the expanded doc block on all the files. This is the first diff that updates js files (except for packager). I'll do follow up diffs for the others.

If you wonder how it's done, by opening every single file and manually changing the header :)

Test Plan: None