[react-packager] Fix regression with transform errors
Summary:
Since we started chaining builds to avoid creating hundreds of concurrent builds when, for example, rebasing. One issue we introduced with that change is that we don't handle build errors correctly and they're indefinitely cached.

This diff uses `finally` and reassigns the package in the cache to keep it up to date.

Test Plan:
* start server
* add syntax error
* refresh page and see redbox
* fix syntax error
* refresh page and error is gone