[AppStateIOS] Fix and simplify AppStateIOS's event subscriptions
The prior implementation of the subscription API was using a function as a key in an object, which I removed. I also simplified the API by removing the unused `type` argument and making the API match that of Facebook's EventEmitter (call subscription.remove() instead of removeEventListener).

Updated the UIExplorer example, which prints out the "active" and "background" states when I background the app in the simulator. No JS errors when I browse around the example and unsubscribe from the event.

There is a bug where the "inactive" state is never hit. This is partly because Apple actually never notifies the app when it is inactive, so we need to listen to UIApplicationWillResignActiveNotification and assume the app will become inactive. Leaving that for another diff if I get around to it or someone else wants to fix it up.