last batch of UIExplorer flowification
Summary:
Finish off @flowifying the UIExplorer source.

Only new wrinkle here is how we're dealing with
`require('image!foo')`: there's a mock module for
each one, in new file ImageMocks.js, which is a
`[lib]` in `.flowconfig`. There's been talk of
adding a wildcarding facility to lighten the
weight of mocking these, but that's out of scope
for f8.

Test Plan: run the samples in UIExplorer