[WIP] Added support for italics and additional font weights
Summary:
This diff adds support for italic fonts to iOS, and also implements proper, CSS-spec compliant font-weight support.

TODO:

Unfortunately, the CSS spec for font weight is not compatible with the previous usage of "bold" in our existing apps:

* In the previous implementation, we hard-coded "bold" to map to fonts with a "-Medium" suffix.
* CSS supports "normal", "bold", "bolder", and "lighter" font weights, but none of these equate to "-Medium" (our current "bold")
* CSS also supports named numeric constants for weight: 100, 200, 300 ... 700, 800, 900
* The numeric value 500 matches "-Medium" (our current "bold"), 400 is normal, 700 is "bold" (the CSS version, not ours)

In the new implementation, I've exactly matched the CSS spec, which also matches the weight constants defined in iOS. I don't think it makes sense to deviate from this. That leaves us with the question of how to fix our code?

I think the simplest solution for now is to codemod all existing use of "bold" to "500", and add an alias in Android that maps "500" to their equivalent of bold. This essentially matches the plan we had before, but instead of the non-spec-compliant "medium" value, we use the spec-compliant "500" value instead.

Moving forward, we probably don't want to have to specify the magic "500" whenever we use bold text, but I'm assuming we would either create a named style for the purpose, or simply specify the font we want to use explicitly. For example, you can specify the font "HelveticaNeue-Medium" either with:

  {
    fontFamily: 'Helvetica Neue',
    fontWeight: '500'
  }

Or, with just:

  {
    fontFamily: 'HelveticaNeue-Medium' // explicit font name
  }

It may seem like a hack to overload "fontFamily" to support explicit font names like this, but I checked in JSFiddle and it's exactly what Safari and Chrome do, so at least we're in good company:

http://jsfiddle.net/410uveuz/

Test Plan: Run the text snapshot tests, once the font changes have been made, and verify they still pass (obviously they don't at the moment, because we haven't code-modded "bold" to "500" yet)