[React Native] Fix CocoaPods spec
Summary: Change 'libicucore' to 'icucore'. The 'libraries' are passed directly to '-l' flags and that expects the 'lib' prefix not to be there.

Test Plan: Create demo project. Add 'React' to Podfile (P19778631). Copy demo files (from 2048 or similar) to project. Create a `UIViewController` with a `RCTRootView` pointing to the local URL. Start the packager (don't forget the `--root` argument to specify the root of your project).