Cancel contents animation before setting new contents
When setting a new image via the imageURL property, the new image
doesn't always replace the previous one when it is finished downloading
because the image view has a previously instated layer animation
on its contents. This cancels the animation prior to setting the new
contents to fix the issue.