[ReactNative] Fix a minor bug in AnimationUtils
Summary:
Some of the easing functions refer to `this` which is undefined due to the way the function is called:

{F22039142}

This fixes that to use `Function.prototype.call` (here is a [[http://stackoverflow.com/a/20375693/90998|good explanation]] of how it works).

Also some minor fixes: `sample` -> `samples` (it's an array), documentation.

Test Plan: In `2048.js`, passed `"easeInOutBounce"` as the easing function, no longer saw the redbox.