Renamed throttleScrollCallbackMS to scrollEventThrottle
Summary: The name throttleScrollCallbackMS was pretty terrible since, a) it's not a callback, and b) the MS is implied. It was also typed as NSUInteger instead of NSTimeInterval, which we now handle automatically. This diff renames it, and also adds some logic to give helpful error messages for anyone using the old version.

Test Plan: Run the apps and tests.