[react-native] Fix race condition in removeSubviewsFromContainerWithID:
Summary:
Previously, this looked at the UI view registry which isn't up to date until the UI blocks execute. Now, it looks at `_shadowViewRegistry` which is updated immediately. The effect of this was that

```
React.render(<Foo />, rootTag);
React.unmountComponentAtNode(rootTag);
```

wouldn't perform the unmount properly, but putting the unmount in a setTimeout would have it work.

Test Plan:
Changed renderApplication.js to have

```
React.unmountComponentAtNode(rootTag);
```

after the React.render call and saw just a white screen in the simulator instead of the app hanging around.