[react-native] Support returning null from a component
Summary:
This previously threw, "Trying to return null from a render, but no null placeholder component was injected."

To avoid circular dependencies, I inlined the parts of React that View was using -- this mirrors what we've done for ReactDOMInput, etc.

Test Plan: Changed render in `AppLauncherRow` to return `null` and saw an empty app launcher.