[ReactNative] RCTView's shadowOffset is of float type, not CGFloat
Summary: D1942342 added a typecheck for each <View>'s properties. The `shadowOffset` prop translates to the `CALayer.shadowOffset` prop, which is `float`, but it was assumed to be `CGFloat`. Apparently this made the value 0.0 even when assigned non-zero value (at least in iPhone 6, iOS 8). The effect is that the <RangeSlider> component "loses" its 2 thumbs, which affects MAdMan age slider. The 2 thumbs are actually still there, but they aren't visible because the shadow is missing.

Test Plan:
Load the RangeSlider example in the catalyst shell. Each slider should have 2 thumbs visible, not "gone".

before:
{F22063649}

after:
{F22063651}