[ReactNative] Normalize name for examples
Summary: One of the biggest difficulty people have when using components or APIs is that we document the props but don't always give a real example that they can just copy and paste. It turns out that we already have those great UIExplorer examples that are working code. I added support for showing those examples in the website using the name as heuristic to find it. It works really well and this diff just renames a couple of files to make them picked up in the docs.

Test Plan:
Run the UIExplorer on catalyst shell, make sure it still works
oss_export and run the website, make sure all those examples show up