[react-native] Update react to 0.13.1, jstransform alongside
Summary:
For https://github.com/facebook/react-native/issues/518. This shouldn't break anything because we only changed a couple of things in betwen the RC and this release.

I also got rid of react-native-github/node_modules/react-tools completely in this diff, because we don't need it and it's the proper way to do what D1952994 meant to do.

Test Plan: Clicked around Madman in Catalyst shell. Also ./runJestTests.sh.