[ReactNative] Allow recover from debugger error
Summary:
The `RCTWebSocketExecutor` returns `nil` when it can't find the chrome debugger.
It was preventing any further reload, since the bridge now will prevent nested
reloads, and the previous reload never ended since there was no context for the
javascript to be executed.

Test Plan:
Launch the app with a invalid URL for the debugger, once it shows the error log
screen try clicking the reload button, it should now work as expected.