[react-packager] Deprecate global image namespace in favor of CommonJS resolution
Summary:
We introduced the `require('image!img')` as a way to disambiguate the image for the JS because images were also globally referred to.
As we move asset packaging to the server this problem should disappear and the JS and image code to live in the same location. Furthermore we would
be able to support requiring images in npm modules. So people publishing components can also publish images that the clients can consume.

The new syntax for require is simply:

`require('./path/to/img.png');`
`require('npmPackage/img.png');`

Test Plan: ./runJestTests.sh