[Exports] Add `invariant` to the exports for Flow compatibility
This is kind of weird but bear with me... Flow special-cases `invariant` so having it available is useful when using Flow. `invariant` is more of a Flow thing than a React thing, but React Native is also setting up the environment with `setTimeout` and `require` so it's kind of a gray area. Alternatively publishing `invariant` to npm instead of using static_upstream's copy would work too.