[ReactNative] rename Animation to AnimationExperimental with warning docs
Summary:
RCTAnimation isn't used internally at all yet, so we don't want to
present it as *the* React Native animation API, so this diff resets
expectations.

I looked at completely pulling it out of OSS, but we have a lot of references to
it and a lot of people have been looking at it, so I'm not sure that's the best
option.  We can still move it out completely after this rename if we want.

Test Plan: 2048 still works.