[ReactNative] Revert D1965911
Summary:
This unbreaks popover. See discussion on D1965911. The initial diff wasn't useful in the first place, we do ensureComponentIsNative on the other Touchable components because we do setNativeProps on them, however TouchableWithoutFeedback doesn't, so there's no need to ensure that.

@sahrens or @frantic do you know how I can write a test so that this would have been caught for next time?

Test Plan: Run madman in the shell app, click the Campaigns popover, make sure it doesn't fatal.